jpeg unpacking with optimization

+++ vsproject/vsproject/utilities.cpp
@@ -441,7 +441,7 @@ int color_space_transform(
         if (rgb2yuv)
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Y
                 tmp[k + red  ] =  0.299f   * img[k + red] + 0.587f   * img[k + green] + 0.114f   * img[k + blue];
@@ -454,7 +454,7 @@ int color_space_transform(
         else
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Red   channel
                 tmp[k + red  ] = img[k + red] + 1.13983f * img[k + blue];
@@ -470,7 +470,7 @@ int color_space_transform(
         if (rgb2yuv)
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Y
                 tmp[k + red  ] =  0.299f * img[k + red] + 0.587f * img[k + green] + 0.114f * img[k + blue];
@@ -483,7 +483,7 @@ int color_space_transform(
         else
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Red   channel
                 tmp[k + red  ] = 1.000f * img[k + red] + 0.000f * img[k + green] + 1.402f * img[k + blue];
@@ -499,7 +499,7 @@ int color_space_transform(
         if (rgb2yuv)
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Y
                 tmp[k + red  ] =  0.333f * img[k + red] + 0.333f * img[k + green] + 0.333f * img[k + blue];
@@ -512,7 +512,7 @@ int color_space_transform(
         else
         {
         #pragma omp parallel for
-            for (int k = 0; k < width * height; k++)
+            for (int k = 0; k < (int)(width * height); k++)
             {
                 //! Red   channel
                 tmp[k + red  ] = 1.0f * img[k + red] + 1.0f * img[k + green] + 0.666f * img[k + blue];
@@ -530,7 +530,7 @@ int color_space_transform(
     }
 
     #pragma omp parallel for
-        for (int k = 0; k < width * height * chnls; k++)
+        for (int k = 0; k < (int)(width * height * chnls); k++)
             img[k] = tmp[k];
 
     return EXIT_SUCCESS;
